+++ jaziku/core/analysis_interval.py
@@ -29,8 +29,8 @@ from jaziku.utils import  console
 class PeriodOfAnalysisInterval(object):
     def __init__(self, type, month=None, day=None):
         self.type = type
-        if type == "trimester":
-            self.trimester = month
+        if type == "trimonthly":
+            self.trimonthly = month
         if type == "15days":
             self.month = month
             self.day = day
@@ -43,10 +43,11 @@ class PeriodOfAnalysisInterval(object):
 
 
 def get_range_analysis_interval():
-    """Return all start days of analysis interval
+    """Return all start days of analysis interval, this is only
+    for analysis interval 5days, 10days and 15days
     """
 
-    if env.config_run.settings['analysis_interval'] != "trimester":
+    if env.config_run.settings['analysis_interval'] not in ['monthly','bimonthly','trimonthly']:
         if env.globals_vars.NUM_DAYS_OF_ANALYSIS_INTERVAL == 5:
             return [1, 6, 11, 16, 21, 26]
         if env.globals_vars.NUM_DAYS_OF_ANALYSIS_INTERVAL == 10:
@@ -57,76 +58,103 @@ def get_range_analysis_interval():
         return None
 
 
-def get_state_of_data():
-    """Calculate and define the state of data based on type of data (daily or monthly)
-    of dependent and independent variable.
-
-    :return:
-        int (1, 2, 3 or 4)
-    """
-    # -------------------------------------------------------------------------
-    # State of the data for process, calculate and write output based on type
-    # of data (daily or monthly) of dependent and independent variable
-    #
-    # | state |  var D  |  var I  |         possible results
-    # |   1   | monthly | monthly |            trimester
-    # |   2   |  daily  | monthly | 5days, 10days, 15days and trimester
-    # |   3   | monthly |  daily  |            trimester
-    # |   4   |  daily  |  daily  | 5days, 10days, 15days and trimester
-    #
-    if env.var_D.is_monthly() and env.var_I.is_monthly():
-        return 1
-    if env.var_D.is_daily() and env.var_I.is_monthly():
-        return 2
-    if env.var_D.is_monthly() and env.var_I.is_daily():
-        return 3
-    if env.var_D.is_daily() and env.var_I.is_daily():
-        return 4
-
-
 def check_analysis_interval():
+    # -------------------------------------------------------------------------
+    # Types of data for process and their results:
+    #
+    # | var D      | var I      | possible results
+    # _____________________________________________________________
+    # | daily      | daily      | 5days, 10days, 15days,
+    # |            | monthly    | monthly, bimonthly*, trimonthly*
+    # |            | bimonthly* |
+    # |            | trimonthly*|
+    # _____________________________________________________________
+    # | monthly    | daily      | monthly, bimonthly*, trimonthly*
+    # |            | monthly    |
+    # |            | bimonthly* |
+    # |            | trimonthly*|
+    # _____________________________________________________________
+    # | bimonthly  | daily      | bimonthly
+    # |            | monthly    |
+    # |            | bimonthly  |
+    # _____________________________________________________________
+    # | trimonthly | daily      | trimonthly
+    # |            | monthly    |
+    # |            | trimonthly |
+    # _____________________________________________________________
+    #  * not combine bimonthly with trimonthly and vice versa
+    #
+    #
+    # Special cases with bimonthly and trimonthly that is
+    # impossible convert the time series, error:
+    #
+    # | var D      | var I      | possible results
+    # ___________________________________________________________
+    # | (any)      | bimonthly  | trimonthly
+    # | (any)      | trimonthly | bimonthly
+    # | bimonthly  | (any)      | trimonthly
+    # | bimonthly  | trimonthly | (any)
+    # ___________________________________________________________
+    #
 
-    if env.globals_vars.STATE_OF_DATA in [1, 3] and env.config_run.settings['analysis_interval'] != "trimester":
-        console.msg_error(_("The var_D of stations have data monthly, but you define\n"
-                            "in runfile the analysis interval as '{0}', this must be,\n"
-                            "in this case, as 'trimester' or use data daily.").format(
-            env.config_run.settings['analysis_interval']))
+    # old state
+    # | state |  var D  |  var I  |         possible results
+    # |   1   | monthly | monthly |            trimonthly
+    # |   2   |  daily  | monthly | 5days, 10days, 15days and trimonthly
+    # |   3   | monthly |  daily  |            trimonthly
+    # |   4   |  daily  |  daily  | 5days, 10days, 15days and trimonthly
+    #
+    # env.globals_vars.STATE_OF_DATA in [1, 3] =  if env.var_D.FREQUENCY_DATA in ['monthly'] and env.var_I.FREQUENCY_DATA in ['daily','monthly']
+    # env.globals_vars.STATE_OF_DATA in [2, 4] =  if env.var_D.FREQUENCY_DATA in ['daily'] and env.var_I.FREQUENCY_DATA in ['daily','monthly']
 
+    def analysis_interval_error(env_variable):
+        console.msg_error(_("The var {0} ({1}) of stations have data {2}, but you\n"
+                            "define in runfile the 'analysis_interval' as '{3}',\n"
+                            "this is incompatible and Jaziku can't convert the times\n"
+                            "series properly for this case. Or change the analysis\n"
+                            "interval or change the data of var {0}.")
+            .format(env.var_.keys()[env.var_.values().index(env_variable)], env_variable.TYPE_SERIES,
+                    env_variable.FREQUENCY_DATA, env.config_run.settings['analysis_interval']))
+
+    if env.var_I.FREQUENCY_DATA in ['bimonthly'] and env.config_run.settings['analysis_interval'] in ["trimonthly"]:
+        analysis_interval_error(env.var_I)
+    if env.var_D.FREQUENCY_DATA in ['bimonthly'] and env.config_run.settings['analysis_interval'] not in ["bimonthly"]:
+        analysis_interval_error(env.var_D)
+    if env.var_I.FREQUENCY_DATA in ['trimonthly'] and env.config_run.settings['analysis_interval'] in ["bimonthly"]:
+        analysis_interval_error(env.var_I)
+    if env.var_D.FREQUENCY_DATA in ['trimonthly'] and env.config_run.settings['analysis_interval'] not in ["trimonthly"]:
+        analysis_interval_error(env.var_D)
+
+    if env.var_D.FREQUENCY_DATA in ['monthly','bimonthly','trimonthly'] and env.config_run.settings['analysis_interval'] not in ["monthly","bimonthly","trimonthly"]:
+        analysis_interval_error(env.var_D)
 
 def adjust_data_of_variables(stations_list):
 
-    if env.globals_vars.STATE_OF_DATA == 3:
-        console.msg(_("   Converting var I of all stations to data monthly ..... "), color='cyan', newline=False)
+    def convert_stations_2(variable, new_freq_data):
+        console.msg(_("   Converting var {0} of all stations to data {1} ..... ")
+                    .format(variable, new_freq_data), color='cyan', newline=False)
         for station in stations_list:
-            station.var_I.daily2monthly()
-            env.var_I.set_FREQUENCY_DATA("monthly", check=False)
-
-        # recalculate global STATE_OF_DATA
-        env.globals_vars.STATE_OF_DATA = get_state_of_data()
+            station.var_[variable].convert2(new_freq_data)
+        env.var_[variable].set_FREQUENCY_DATA(new_freq_data, check=False)
         console.msg(_("done"), color='green')
 
+    if env.config_run.settings['analysis_interval'] == "monthly":
+        for variable in ['D', 'I']:
+            if env.var_[variable].FREQUENCY_DATA in ['daily']:
+                convert_stations_2(variable, 'monthly')
         return True
 
-    if env.globals_vars.STATE_OF_DATA in [2, 4]:
-        # if analysis_interval is defined by trimester but var_I or/and var_D has data
-        # daily, first convert in data monthly and continue with results by trimester
-        if env.config_run.settings['analysis_interval'] == "trimester":
-            console.msg(_("   Converting var D of all stations to data monthly ..... "), color='cyan', newline=False)
-            for station in stations_list:
-                station.var_D.daily2monthly()
-                env.var_D.set_FREQUENCY_DATA("monthly", check=False)
-            console.msg(_("done"), color='green')
-            if env.globals_vars.STATE_OF_DATA == 4:
-                console.msg(_("   Converting var I of all stations to data monthly ..... "), color='cyan', newline=False)
-                for station in stations_list:
-                    station.var_I.daily2monthly()
-                    env.var_I.set_FREQUENCY_DATA("monthly", check=False)
-                console.msg(_("done"), color='green')
+    if env.config_run.settings['analysis_interval'] == "bimonthly":
+        for variable in ['D', 'I']:
+            if env.var_[variable].FREQUENCY_DATA in ['daily','monthly']:
+                convert_stations_2(variable, 'bimonthly')
+        return True
 
-            # recalculate global STATE_OF_DATA
-            env.globals_vars.STATE_OF_DATA = get_state_of_data()
-
-            return True
+    if env.config_run.settings['analysis_interval'] == "trimonthly":
+        for variable in ['D', 'I']:
+            if env.var_[variable].FREQUENCY_DATA in ['daily','monthly']:
+                convert_stations_2(variable, 'trimonthly')
+        return True
 
     return False
 
@@ -142,22 +170,22 @@ def locate_day_in_analysis_interval(day_for_locate):
             return range_item
 
 
-def get_values_in_range_analysis_interval(station, type, year, month, day=None, lag=None):
+def get_values_in_range_analysis_interval(variable, year, n_month, day=None, lag=None):
     """Get all values inside range analysis interval in specific year, month, day or lag.
-    The "type" must be "D" or "I". For trimester the "month" is the start month of
-    trimester, for 5, 10, or 15 days the "day" is a start day in the range analysis
-    interval. The "lag" is only affect for the independent variable and it must
-    be 0, 1 or 2.
+    The "type" must be "D" or "I". For bimonthly and trimonthly the "n_month" is the start
+    of month of bimonthly or trimonthly respective, for 5, 10, or 15 days the "day" is a
+    start day in the range analysis interval. The "lag" is only affect for the independent
+    variable and it must be 0, 1 or 2.
     """
 
     range_analysis_interval = get_range_analysis_interval()
 
     # first fix if day not is a valid start day in analysis interval
-    if day and  range_analysis_interval:
+    if day and range_analysis_interval:
         if day not in range_analysis_interval:
             day = locate_day_in_analysis_interval(day)
 
-    if type == 'D':
+    if variable.type == 'D':
         var_D_values = []
         if env.var_D.is_daily():
             # clone range for add the last day (32) for calculate interval_day_var_D
@@ -168,20 +196,17 @@ def get_values_in_range_analysis_interval(station, type, year, month, day=None, 
             interval_day_var_D = range(day - 1, rai_plus[rai_plus.index(day) + 1] - 1)
 
             for iter_day in interval_day_var_D:
-                now = date(year, month, 1) + relativedelta(days=iter_day)
+                now = date(year, n_month, 1) + relativedelta(days=iter_day)
                 # check if continues with the same month
-                if now.month == month:
-                    index_var_D = station.var_D.date.index(now)
-                    var_D_values.append(station.var_D.data[index_var_D])
-        if env.var_D.is_monthly():
-            # get the three values for var_D in this month
-            for iter_month in range(3):
-                var_D_values.append(station.var_D.data[station.var_D.date.index(
-                    date(year, month, 1) + relativedelta(months=iter_month))])
+                if now.month == n_month:
+                    index_var_D = variable.date.index(now)
+                    var_D_values.append(variable.data[index_var_D])
+        if env.var_D.is_n_monthly():
+            var_D_values.append(variable.data[variable.date.index(date(year, n_month, 1))])
 
         return var_D_values
 
-    if type == 'I':
+    if variable.type == 'I':
         var_I_values = []
         if env.var_I.is_daily():
             # from day to next iterator based on analysis interval
@@ -191,7 +216,7 @@ def get_values_in_range_analysis_interval(station, type, year, month, day=None, 
             except:
                 end_interval = range_analysis_interval[0]
 
-            start_date = date(year, month, start_interval)
+            start_date = date(year, n_month, start_interval)
 
             if range_analysis_interval.index(day) - lag < 0:
                 start_date += relativedelta(months= -1)
@@ -199,29 +224,22 @@ def get_values_in_range_analysis_interval(station, type, year, month, day=None, 
             iter_date = start_date
 
             while iter_date.day != end_interval:
-                index_var_I = station.var_I.date.index(iter_date)
-                var_I_values.append(station.var_I.data[index_var_I])
+                index_var_I = variable.date.index(iter_date)
+                var_I_values.append(variable.data[index_var_I])
                 iter_date += relativedelta(days=1)
 
-        if env.var_I.is_monthly():
-            if env.globals_vars.STATE_OF_DATA in [1, 3]:
-                # get the three values for var_I in this month
-                for iter_month in range(3):
-                    var_I_values.append(station.var_I.data[station.var_I.date.index(
-                        date(year, month, 1) + relativedelta(months=iter_month - lag))])
-            if env.globals_vars.STATE_OF_DATA in [2]:
-                # keep constant value for month
-                if env.config_run.settings['analysis_interval'] == "trimester":
-                    var_I_values.append(station.var_I.data[station.var_I.date.index(
-                        date(year, month, 1) + relativedelta(months= -lag))])
-                else:
-                    real_date = date(year, month, day) + relativedelta(days= -env.globals_vars.NUM_DAYS_OF_ANALYSIS_INTERVAL * lag)
-                    # e.g if lag 2 in march and calculate to 15days go to february and not january
-                    if month - real_date.month > 1:
-                        real_date = date(real_date.year, real_date.month + 1, 1)
+        if env.var_I.is_n_monthly():
+            if env.config_run.settings['analysis_interval'] in ["5days", "10days", "15days"]:
+                real_date = date(year, n_month, day) + relativedelta(days= -env.globals_vars.NUM_DAYS_OF_ANALYSIS_INTERVAL * lag)
+                # e.g if lag 2 in march and calculate to 15days go to february and not january
+                if n_month - real_date.month > 1:
+                    real_date = date(real_date.year, real_date.month + 1, 1)
 
-                    var_I_values.append(station.var_I.data[station.var_I.date.index(
-                        date(real_date.year, real_date.month, 1))])
+                var_I_values.append(variable.data[variable.date.index(
+                    date(real_date.year, real_date.month, 1))])
+            else:
+                var_I_values.append(variable.data[variable.date.index(
+                        date(year, n_month, 1) + relativedelta(months=-lag))])
 
         return var_I_values