+++ jaziku/utils/output.py
@@ -35,16 +35,7 @@ def number(num, accuracy=False):
     else:
         return str(round(float(num), env.globals_vars.ACCURACY)).replace('.', ',')
 
-def trimester_in_initials(trimester):
-    """
-    Return the three first letters of three months of trimester
-    """
-    _trim_text = {-2: _('NDJ'), -1: _('DJF'), 0: _('JFM'), 1: _('FMA'), 2: _('MAM'),
-                 3: _('AMJ'), 4: _('MJJ'), 5: _('JJA'), 6: _('JAS'), 7: _('ASO'),
-                 8: _('SON'), 9: _('OND'), 10: _('NDJ'), 11: _('DJF')}
-    return _trim_text[trimester]
-
-def month_in_initials(month):
+def months_in_initials(month):
     """
     Return the initials of month (first three letters)
     """
@@ -52,6 +43,52 @@ def month_in_initials(month):
                   3: _('Apr'), 4: _('May'), 5: _('Jun'), 6: _('Jul'), 7: _('Aug'),
                   8: _('Sep'), 9: _('Oct'), 10: _('Nov'), 11: _('Dec')}
     return _month_text[month]
+
+def bi_months_in_initials(bimonthly):
+    """
+    Return the three first letters of two months of bimonthly
+    """
+    _bim_text = {-2: _('NovDic'), -1: _('DicJan'), 0: _('JanFeb'), 1: _('FebMar'), 2: _('MarApr'),
+                 3: _('AprMay'), 4: _('MayJun'), 5: _('JunJul'), 6: _('JulAug'), 7: _('AugSep'),
+                  8: _('SepOct'), 9: _('OctNov'), 10: _('NovDic'), 11: _('DecJan')}
+    return _bim_text[bimonthly]
+
+def tri_months_in_initials(trimonthly):
+    """
+    Return the three first letters of three months of trimonthly
+    """
+    _trim_text = {-2: _('NDJ'), -1: _('DJF'), 0: _('JFM'), 1: _('FMA'), 2: _('MAM'),
+                 3: _('AMJ'), 4: _('MJJ'), 5: _('JJA'), 6: _('JAS'), 7: _('ASO'),
+                 8: _('SON'), 9: _('OND'), 10: _('NDJ'), 11: _('DJF')}
+    return _trim_text[trimonthly]
+
+def n_months_in_initials(type, n_month):
+    if env.var_[type].is_monthly():
+        return months_in_initials(n_month)
+    if env.var_[type].is_bimonthly():
+        return bi_months_in_initials(n_month)
+    if env.var_[type].is_trimonthly():
+        return tri_months_in_initials(n_month)
+
+def n_monthly_int2char(int_month, type=False, n_monthly=False):
+    int_month = int(int_month)
+    if (type and env.var_[type].is_monthly()) or n_monthly == 1:
+        return fix_zeros(int_month)
+    if (type and env.var_[type].is_bimonthly()) or n_monthly == 2:
+        char_month_dict = {1:'jf', 2: 'fm', 3: 'ma',4: 'am', 5: 'mj', 6: 'jj', 7: 'ja', 8: 'as', 9: 'so', 10: 'on', 11: 'nd', 12: 'dj'}
+    if (type and env.var_[type].is_trimonthly()) or n_monthly == 3:
+        char_month_dict = {1:'jfm', 2: 'fma', 3: 'mam',4: 'amj', 5: 'mjj', 6: 'jja', 7: 'jas', 8: 'aso', 9: 'son', 10: 'ond', 11: 'ndj', 12: 'djf'}
+
+    if int_month in char_month_dict:
+        return char_month_dict[int_month]
+    else:
+        return False
+
+def bimonthly_int2char(int_month):
+    return n_monthly_int2char(int_month, n_monthly=2)
+
+def trimonthly_int2char(int_month):
+    return n_monthly_int2char(int_month, n_monthly=3)
 
 def make_dirs(_dir):
     if not os.path.isdir(_dir):